requirements.txt
streamlit==1.41.1
google-genai==1.0.0
pymupdf==1.25.0
pandas==2.2.0
python-dotenv==1.0.0
openai==1.50.0
requests==2.31.0
tenacity==8.2.3
pytest==8.0.0

